# apps/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ProblemDetail": {
        "description": "RFC 9457 problem-details body - the API's single error contract.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemFieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
      "ProblemFieldError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ProblemFieldError",
        "type": "object"
      },
...
      "ScenarioCreate": {
        "properties": {
          "attacking_team_id": {
            "default": "england",
            "title": "Attacking Team Id",
            "type": "string"
          },
          "defending_team_id": {
            "default": "argentina",
            "title": "Defending Team Id",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "routine_id": {
            "title": "Routine Id",
            "type": "string"
          },
          "scheme_id": {
            "title": "Scheme Id",
            "type": "string"
          }
        },
        "required": [
          "name",
          "routine_id",
          "scheme_id"
        ],
        "title": "ScenarioCreate",
        "type": "object"
      },
      "ScenarioDTO": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "scenario_hash": {
            "title": "Scenario Hash",
            "type": "string"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "spec": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Spec",
            "type": "object"
          }
        },
        "required": [
          "scenario_id",
          "name",
          "spec",
          "scenario_hash",
          "created_at"
        ],
        "title": "ScenarioDTO",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "summary": "Set-piece simulation, Monte Carlo, and scenario API for Restart Lab.",
    "title": "Restart Lab API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scenarios": {
      "get": {
        "operationId": "list_scenarios_api_v1_scenarios_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioDTO"
                  },
                  "title": "Response List Scenarios Api V1 Scenarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Rate limit exceeded"
          }
        },
        "summary": "List Scenarios",
        "tags": [
          "scenarios"
        ]
      },
      "post": {
        "operationId": "create_scenario_api_v1_scenarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Scenario",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/api/v1/scenarios/{scenario_id}": {
      "get": {
        "operationId": "get_scenario_api_v1_scenarios__scenario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Get Scenario",
        "tags": [
          "scenarios"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "This deployment",
      "url": "/"
    }
  ]
}